+++ packages/ui/src/components/Button/Button.css.ts
@@ -81,7 +81,24 @@ export const buttonRecipe = recipe({
           },
         },
       },
-      line: {},
+      line: {
+        border: '0.2rem solid transparent',
+        backgroundClip: 'padding-box',
+        backgroundColor: tokens.colors.purple100,
+        color: tokens.colors.primary800,
+
+        selectors: {
+          '&:hover': {
+            color: tokens.colors.primary800,
+            backgroundColor: tokens.colors.purple100,
+          },
+          '&:disabled': {
+            backgroundColor: tokens.colors.grey0,
+            color: tokens.colors.grey200,
+            borderColor: tokens.colors.grey100,
+          },
+        },
+      },
     },
 
     isLoading: {
