+++ packages/ui/src/components/Button/Button.tsx
@@ -6,6 +6,7 @@ import {
   spinner,
 } from './Button.css';
 import { Spinner, SpinnerProps } from '../Spinner';
+import { LineButtonAnimate } from './LineButtonAnimate';
 
 export type ButtonSize = 'small' | 'large';
 export type ButtonVariant = 'primary' | 'neutral' | 'text' | 'line';
@@ -38,26 +39,35 @@ export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
       ...rest
     },
     ref
-  ) => (
-    <button
-      ref={ref}
-      className={`${buttonRecipe({ size, variant, isLoading })} ${className}`}
-      {...rest}
-    >
-      {isLoading && (
-        <Spinner
-          className={spinner({ size, isLoading })}
-          color={SpinnerColor[variant]}
-          size={size}
-        />
-      )}
-      <div className={`${buttonChildrenRecipe({ size, isLoading })}`}>
-        {leftAddon && <div className={addonRootStyle[size]}>{leftAddon}</div>}
-        {children}
-        {rightAddon && <div className={addonRootStyle[size]}>{rightAddon}</div>}
-      </div>
-    </button>
-  )
+  ) => {
+    const content = (
+      <button
+        ref={ref}
+        className={`${buttonRecipe({ size, variant, isLoading })} ${className}`}
+        {...rest}
+      >
+        {isLoading && (
+          <Spinner
+            className={spinner({ size, isLoading })}
+            color={SpinnerColor[variant]}
+            size={size}
+          />
+        )}
+        <div className={`${buttonChildrenRecipe({ size, isLoading })}`}>
+          {leftAddon && <div className={addonRootStyle[size]}>{leftAddon}</div>}
+          {children}
+          {rightAddon && (
+            <div className={addonRootStyle[size]}>{rightAddon}</div>
+          )}
+        </div>
+      </button>
+    );
+
+    if (variant === 'line') {
+      return <LineButtonAnimate size={size}>{content}</LineButtonAnimate>;
+    }
+    return <>{content}</>;
+  }
 );
 
 Button.displayName = 'Button';
