--- packages/ui/src/components/LineButton/index.ts
@@ -1,1 +1,0 @@
-export { LineButton } from './LineButton';
